requirements.txt
faiss-cpu
torch
numpy
streamlit
transformers
llama-cpp-python
PyPDF2
python-docx
requests

